Agrega: Reserva de entradas
 - Renombra datatype y remueve servicio de ejemplo.

 - Se renombra el datatype DataDisponibilidad a DataHorario, dado que se mapea mejor con la realidad.
 - Tambien se remueve el servicio de Hello World ya que tenemos uno de verdad.

 - DataHorario tendra DataDisponibilidad como atributo en vez de la disponibilidad

 - Agrega emf util, cron y algo de realizar reserva

 - Remueve cron por reserva y agrega un unico cron para anular.
 Este commit existe con el fin de hacer un buen uso de los recursos;

 - Finaliza el servicio de reservar:
  - Se cambia long por id en eventos, ya que la letra asi lo indica
  - Los datatypes deben de implementar Serializable y tener setters
  - Se agrega fecha de creacion y actualizacion para las reservas

Agrega trigger de postgres para calcular la cantidad de disponibilidades

+++ lab-2/ticketinco-soap/src/main/java/ticketinco/dao/EventoDAOJpa.java
@@ -18,11 +18,11 @@ public class EventoDAOJpa<T> extends GenericJpaDAO<T, Serializable> implements I
 
     public  List<Evento> getEventos() {return null;}
 
-    public Evento getEventoById (long id) {
+    public Evento getEventoById (int id) {
         return (Evento) this.read(id);
     }
 
-    public Evento getEventoByIdAndDate (long id, Date fecha) {
+    public Evento getEventoByIdAndDate (int id, Date fecha) {
         return (Evento) this.read(id);
     }
 }
